Introduce new async getHttpContent method
