+++ http/org.wso2.carbon.transport.http.netty/src/main/java/org/wso2/carbon/transport/http/netty/message/MessageListener.java
@@ -1,0 +1,11 @@
+package org.wso2.carbon.transport.http.netty.message;
+
+import io.netty.handler.codec.http.HttpContent;
+
+/**
+ * Get notified when there is a state change in message
+ */
+public interface MessageListener {
+
+    void onMessage(HttpContent httpContent);
+}
