+++ http/org.wso2.carbon.transport.http.netty/src/main/java/org/wso2/carbon/transport/http/netty/sender/channel/pool/ConnectionManager.java
@@ -98,8 +98,8 @@ public class ConnectionManager {
             EventLoopGroup group;
             ChannelHandlerContext ctx = sourceHandler.getInboundChannelContext();
             // TODO: use the same event loop group once pass-through service detection is complete.
-//            group = ctx.channel().eventLoop();
-            group = targetEventLoopGroup;
+            group = ctx.channel().eventLoop();
+//            group = targetEventLoopGroup;
             Class cl = ctx.channel().getClass();
 
             if (poolManagementPolicy == PoolManagementPolicy.LOCK_DEFAULT_POOLING) {
