+++ http/org.wso2.carbon.transport.http.netty/src/test/java/org/wso2/carbon/transport/http/netty/contentaware/serverConnectorListeners/EchoMessageListener.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.wso2.carbon.transport.http.netty.contentaware;
+package org.wso2.carbon.transport.http.netty.contentaware.serverConnectorListeners;
 
 import io.netty.handler.codec.http.HttpHeaders;
 import org.slf4j.Logger;
