+++ http/org.wso2.carbon.transport.http.netty/src/test/java/org/wso2/carbon/transport/http/netty/contentaware/serverConnectorListeners/RequestResponseCreationStreamingListener.java
@@ -17,10 +17,9 @@
  *
  */
 
-package org.wso2.carbon.transport.http.netty.contentaware;
+package org.wso2.carbon.transport.http.netty.contentaware.serverConnectorListeners;
 
-import io.netty.buffer.Unpooled;
-import io.netty.handler.codec.http.DefaultLastHttpContent;
+import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.carbon.transport.http.netty.common.Constants;
@@ -35,12 +34,13 @@ import org.wso2.carbon.transport.http.netty.contract.ServerConnectorException;
 import org.wso2.carbon.transport.http.netty.contractimpl.HttpWsConnectorFactoryImpl;
 import org.wso2.carbon.transport.http.netty.message.HTTPCarbonMessage;
 import org.wso2.carbon.transport.http.netty.message.HTTPConnectorUtil;
+import org.wso2.carbon.transport.http.netty.message.HttpMessageDataStreamer;
 import org.wso2.carbon.transport.http.netty.util.TestUtil;
 
-import java.io.UnsupportedEncodingException;
-import java.nio.ByteBuffer;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ExecutorService;
@@ -48,17 +48,16 @@ import java.util.concurrent.Executors;
 import java.util.stream.Collectors;
 
 /**
- * A Message Processor which creates Request and Response
+ * A class which read and write content through streams
  */
-public class RequestResponseCreationListener implements HttpConnectorListener {
-    private Logger logger = LoggerFactory.getLogger(RequestResponseCreationListener.class);
+public class RequestResponseCreationStreamingListener implements HttpConnectorListener {
 
-    private String responseValue;
+    private static final Logger logger = LoggerFactory.getLogger(RequestResponseCreationStreamingListener.class);
+
     private TransportsConfiguration configuration;
     private ExecutorService executor = Executors.newSingleThreadExecutor();
 
-    public RequestResponseCreationListener(String responseValue, TransportsConfiguration configuration) {
-        this.responseValue = responseValue;
+    public RequestResponseCreationStreamingListener(TransportsConfiguration configuration) {
         this.configuration = configuration;
     }
 
@@ -66,21 +65,15 @@ public class RequestResponseCreationListener implements HttpConnectorListener {
     public void onMessage(HTTPCarbonMessage httpRequest) {
         executor.execute(() -> {
             try {
-                int length = httpRequest.getFullMessageLength();
-                List<ByteBuffer> byteBufferList = httpRequest.getFullMessageBody();
-                ByteBuffer byteBuffer = ByteBuffer.allocate(length);
-                byteBufferList.forEach(byteBuffer::put);
-                String requestValue = new String(byteBuffer.array());
-                byte[] arry = responseValue.getBytes("UTF-8");
+                HttpMessageDataStreamer streamer = new HttpMessageDataStreamer(httpRequest);
+                InputStream inputStream = streamer.getInputStream();
 
                 HTTPCarbonMessage newMsg = httpRequest.cloneCarbonMessageWithOutData();
-                if (newMsg.getHeader(Constants.HTTP_TRANSFER_ENCODING) == null) {
-                    newMsg.setHeader(Constants.HTTP_CONTENT_LENGTH, String.valueOf(arry.length));
-                }
-                ByteBuffer byteBuffer1 = ByteBuffer.allocate(arry.length);
-                byteBuffer1.put(arry);
-                byteBuffer1.flip();
-                newMsg.addHttpContent(new DefaultLastHttpContent(Unpooled.wrappedBuffer(byteBuffer1)));
+                OutputStream outputStream = new HttpMessageDataStreamer(newMsg).getOutputStream();
+                byte[] bytes = IOUtils.toByteArray(inputStream);
+                outputStream.write(bytes);
+                outputStream.flush();
+                outputStream.close();
                 newMsg.setProperty(Constants.HOST, TestUtil.TEST_HOST);
                 newMsg.setProperty(Constants.PORT, TestUtil.TEST_HTTP_SERVER_PORT);
 
@@ -99,39 +92,26 @@ public class RequestResponseCreationListener implements HttpConnectorListener {
                 HttpWsConnectorFactory httpWsConnectorFactory = new HttpWsConnectorFactoryImpl();
                 HttpClientConnector clientConnector =
                         httpWsConnectorFactory.createHttpClientConnector(transportProperties, senderConfiguration);
-
                 HttpResponseFuture future = clientConnector.send(newMsg);
                 future.setHttpConnectorListener(new HttpConnectorListener() {
                     @Override
-                    public void onMessage(HTTPCarbonMessage httpResponse) {
+                    public void onMessage(HTTPCarbonMessage httpMessage) {
                         executor.execute(() -> {
-                            int length = httpResponse.getFullMessageLength();
-                            List<ByteBuffer> byteBufferList = httpResponse.getFullMessageBody();
+                            HttpMessageDataStreamer streamer = new HttpMessageDataStreamer(httpMessage);
+                            InputStream inputStream = streamer.getInputStream();
 
-                            ByteBuffer byteBuffer = ByteBuffer.allocate(length);
-                            byteBufferList.forEach(byteBuffer::put);
-                            String responseValue = new String(byteBuffer.array()) + ":" + requestValue;
-                            byte[] array = new byte[0];
+                            HTTPCarbonMessage newMsg = httpMessage.cloneCarbonMessageWithOutData();
+                            OutputStream outputStream = new HttpMessageDataStreamer(newMsg).getOutputStream();
                             try {
-                                array = responseValue.getBytes("UTF-8");
-                            } catch (UnsupportedEncodingException e) {
-
+                                byte[] bytes = IOUtils.toByteArray(inputStream);
+                                outputStream.write(bytes);
+                                outputStream.flush();
+                                outputStream.close();
+                            } catch (IOException e) {
+                                throw new RuntimeException("Cannot read Input Stream from Response", e);
                             }
-
-                            ByteBuffer byteBuff = ByteBuffer.allocate(array.length);
-                            byteBuff.put(array);
-                            byteBuff.flip();
-                            HTTPCarbonMessage httpCarbonMessage = httpResponse
-                                    .cloneCarbonMessageWithOutData();
-                            if (httpCarbonMessage.getHeader(Constants.HTTP_TRANSFER_ENCODING) == null) {
-                                httpCarbonMessage.setHeader(Constants.HTTP_CONTENT_LENGTH,
-                                        String.valueOf(array.length));
-                            }
-                            httpCarbonMessage.addHttpContent(
-                                    new DefaultLastHttpContent(Unpooled.wrappedBuffer(byteBuff)));
-
                             try {
-                                httpRequest.respond(httpCarbonMessage);
+                                httpRequest.respond(newMsg);
                             } catch (ServerConnectorException e) {
                                 logger.error("Error occurred during message notification: " + e.getMessage());
                             }
@@ -143,13 +123,11 @@ public class RequestResponseCreationListener implements HttpConnectorListener {
 
                     }
                 });
-            } catch (UnsupportedEncodingException e) {
-                logger.error("Encoding is not supported", e);
+
             } catch (Exception e) {
-                logger.error("Failed to send the message to the back-end", e);
+                logger.error("Error while reading stream", e);
             }
         });
-
     }
 
     @Override
