+++ http/org.wso2.carbon.transport.http.netty/src/test/java/org/wso2/carbon/transport/http/netty/contentaware/serverConnectorListeners/RequestResponseTransformStreamingListener.java
@@ -17,10 +17,9 @@
  *
  */
 
-package org.wso2.carbon.transport.http.netty.contentaware;
+package org.wso2.carbon.transport.http.netty.contentaware.serverConnectorListeners;
 
-import io.netty.buffer.Unpooled;
-import io.netty.handler.codec.http.DefaultLastHttpContent;
+import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.carbon.transport.http.netty.common.Constants;
@@ -35,12 +34,13 @@ import org.wso2.carbon.transport.http.netty.contract.ServerConnectorException;
 import org.wso2.carbon.transport.http.netty.contractimpl.HttpWsConnectorFactoryImpl;
 import org.wso2.carbon.transport.http.netty.message.HTTPCarbonMessage;
 import org.wso2.carbon.transport.http.netty.message.HTTPConnectorUtil;
+import org.wso2.carbon.transport.http.netty.message.HttpMessageDataStreamer;
 import org.wso2.carbon.transport.http.netty.util.TestUtil;
 
-import java.io.UnsupportedEncodingException;
-import java.nio.ByteBuffer;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ExecutorService;
@@ -48,44 +48,29 @@ import java.util.concurrent.Executors;
 import java.util.stream.Collectors;
 
 /**
- * Transform message in request and response path
+ * Streaming processor which reads from same and write to same message
  */
-public class RequestResponseTransformListener implements HttpConnectorListener {
+public class RequestResponseTransformStreamingListener implements HttpConnectorListener {
 
-    private static final Logger logger = LoggerFactory.getLogger(RequestResponseTransformListener.class);
-
+    private static final Logger logger = LoggerFactory.getLogger(RequestResponseTransformStreamingListener.class);
     private ExecutorService executor = Executors.newSingleThreadExecutor();
-    private String responseValue;
-    private String requestValue;
     private TransportsConfiguration configuration;
 
-    public RequestResponseTransformListener(String responseValue, TransportsConfiguration configuration) {
-        this.responseValue = responseValue;
+    public RequestResponseTransformStreamingListener(TransportsConfiguration configuration) {
         this.configuration = configuration;
     }
 
     @Override
-    public void onMessage(HTTPCarbonMessage httpRequest) {
+    public void onMessage(HTTPCarbonMessage httpRequestMessage) {
         executor.execute(() -> {
             try {
-                int length = httpRequest.getFullMessageLength();
-                List<ByteBuffer> byteBufferList = httpRequest.getFullMessageBody();
-
-                ByteBuffer byteBuff = ByteBuffer.allocate(length);
-                byteBufferList.forEach(byteBuff::put);
-                requestValue = new String(byteBuff.array());
-
-                httpRequest.setProperty(Constants.HOST, TestUtil.TEST_HOST);
-                httpRequest.setProperty(Constants.PORT, TestUtil.TEST_HTTP_SERVER_PORT);
-
-                if (responseValue != null) {
-                    byte[] array = responseValue.getBytes("UTF-8");
-                    ByteBuffer byteBuffer = ByteBuffer.allocate(array.length);
-                    byteBuffer.put(array);
-                    httpRequest.setHeader(Constants.HTTP_CONTENT_LENGTH, String.valueOf(array.length));
-                    byteBuffer.flip();
-                    httpRequest.addHttpContent(new DefaultLastHttpContent(Unpooled.wrappedBuffer(byteBuffer)));
-                }
+                InputStream inputStream = new HttpMessageDataStreamer(httpRequestMessage).getInputStream();
+                OutputStream outputStream = new HttpMessageDataStreamer(httpRequestMessage).getOutputStream();
+                byte[] bytes = IOUtils.toByteArray(inputStream);
+                outputStream.write(bytes);
+                outputStream.close();
+                httpRequestMessage.setProperty(Constants.HOST, TestUtil.TEST_HOST);
+                httpRequestMessage.setProperty(Constants.PORT, TestUtil.TEST_HTTP_SERVER_PORT);
 
                 Map<String, Object> transportProperties = new HashMap<>();
                 Set<TransportProperty> transportPropertiesSet = configuration.getTransportProperties();
@@ -95,42 +80,31 @@ public class RequestResponseTransformListener implements HttpConnectorListener {
 
                 }
 
-                String scheme = (String) httpRequest.getProperty(Constants.PROTOCOL);
+                String scheme = (String) httpRequestMessage.getProperty(Constants.PROTOCOL);
                 SenderConfiguration senderConfiguration = HTTPConnectorUtil
                         .getSenderConfiguration(configuration, scheme);
 
                 HttpWsConnectorFactory httpWsConnectorFactory = new HttpWsConnectorFactoryImpl();
                 HttpClientConnector clientConnector =
                         httpWsConnectorFactory.createHttpClientConnector(transportProperties, senderConfiguration);
-                HttpResponseFuture future = clientConnector.send(httpRequest);
+                HttpResponseFuture future = clientConnector.send(httpRequestMessage);
                 future.setHttpConnectorListener(new HttpConnectorListener() {
                     @Override
-                    public void onMessage(HTTPCarbonMessage httpMessage) {
+                    public void onMessage(HTTPCarbonMessage httpResponse) {
                         executor.execute(() -> {
-                            int length = httpMessage.getFullMessageLength();
-                            List<ByteBuffer> byteBufferList = httpMessage.getFullMessageBody();
-
-                            ByteBuffer byteBuffer = ByteBuffer.allocate(length);
-                            byteBufferList.forEach(byteBuffer::put);
-                            String responseValue = new String(byteBuffer.array()) + ":" + requestValue;
-                            if (requestValue != null) {
-                                byte[] array = new byte[0];
-                                try {
-                                    array = responseValue.getBytes("UTF-8");
-                                } catch (UnsupportedEncodingException e) {
-
-                                }
-                                ByteBuffer byteBuff = ByteBuffer.allocate(array.length);
-                                byteBuff.put(array);
-                                httpMessage.setHeader(Constants.HTTP_CONTENT_LENGTH, String.valueOf(array.length));
-                                byteBuff.flip();
-                                httpMessage.addHttpContent(
-                                        new DefaultLastHttpContent(Unpooled.wrappedBuffer(byteBuff)));
-                                try {
-                                    httpRequest.respond(httpMessage);
-                                } catch (ServerConnectorException e) {
-                                    logger.error("Error occurred during message notification: " + e.getMessage());
-                                }
+                            InputStream inputS = new HttpMessageDataStreamer(httpResponse).getInputStream();
+                            OutputStream outputS = new HttpMessageDataStreamer(httpResponse).getOutputStream();
+                            try {
+                                byte[] bytes = IOUtils.toByteArray(inputS);
+                                outputS.write(bytes);
+                                outputS.close();
+                            } catch (IOException e) {
+                                throw new RuntimeException("Cannot read Input Stream from Response", e);
+                            }
+                            try {
+                                httpRequestMessage.respond(httpResponse);
+                            } catch (ServerConnectorException e) {
+                                logger.error("Error occurred during message notification: " + e.getMessage());
                             }
                         });
                     }
